derive the true damage/coin totals from a server-side replay, not the client's claim
Builds the actual verification mechanism the last two commits' event log
was collected for. classes/local/engine/replay.php re-simulates a phase
from its recorded seed and event log: a turn loop alternates player and
boss, resolving each swap's full cascade through board_engine.php/
combat_engine.php exactly like board.js::checkMatches() does, applying a
'move' event's real board outcome and a 'question' event's already-known
HP/multiplier effect at the precise point in the sequence the replay's own
simulation determines it belongs — nothing about the mana-crossing timing
needs to come from outside the simulation itself, since resolve_match_effects
detects it identically to the live client.

The whole thing only ever returns a value once it reaches a genuine
terminal state, boss dead for good or player dead. Running out of
recorded events with neither side dead is deliberately never treated as a
partial answer — it's exactly the kind of gap (a lost checkpoint, a
sendBeacon that never landed) the fail-safe policy exists to absorb
without ever crediting less than a match that actually finished earned.
Same treatment for a caught exception, an engine-version mismatch, or a
cascade that never settles within a generous round cap.

classes/local/replay_credit.php sits between this and save_progress.php/
advance_phase.php: when the replay reaches a conclusive result, that's
what gets persisted and banked, firing a replay_diverged event first if
it disagrees with the client's own claim (observability only — the
derived value is credited either way, never the claim). When it can't,
the claim passes through completely unchanged, and every existing sanity
clamp (the boss-HP cap, the coin ceiling) still applies exactly as before.

Two real bugs turned up while wiring this together, neither caught by any
static tool. combat_engine::resolve_damage()/resolve_poison_tick() had
$hp typed int since the engine was first ported — latent until a
multi-round simulation actually produced fractional HP (a Potion heals
baseDamage/4, a question crit is baseDamage*3*a tenths-precision
multiplier) and silent truncation started pulling it away from the JS
side, which never rounds between hits. Fixed to float. And the match
groups board_engine reports ({type, cells}) were never reshaped into what
combat_engine expects ({type, size}) before being passed through — the
same reshape board.js does before calling combat_rules.js — so every
combo multiplier silently came out NaN. Only surfaced once a real
scenario, generated by running the actual board_rules.js headless through
Node, was pushed through the whole loop instead of the pure functions in
isolation.

+++ classes/local/engine/combat_engine.php
@@ -159,17 +159,20 @@ class combat_engine {
      * Resolves one side taking damage: a ready shield blocks the hit entirely instead of
      * reducing HP.
      *
-     * @param int $hp Current HP of the side being hit.
+     * @param float $hp Current HP of the side being hit — a float, not an int: repeated
+     *  fractional damage/heal amounts (a Potion's baseDamage / 4, a crit's baseDamage * 3 *
+     *  a multiplier ending in a tenth) accumulate across a whole phase without ever being
+     *  rounded back to a whole number mid-match, exactly like the JS side's untyped number.
      * @param bool $shieldready Whether that side's shield is currently armed.
      * @param float $amount Damage amount before the shield check.
      * @return array ['newHp' => float, 'shieldConsumed' => bool, 'appliedAmount' => float]
      */
-    public static function resolve_damage(int $hp, bool $shieldready, float $amount): array {
+    public static function resolve_damage(float $hp, bool $shieldready, float $amount): array {
         $appliedamount = $amount;
         $shieldconsumed = false;
         if ($shieldready) {
             $shieldconsumed = true;
-            $appliedamount = 0;
+            $appliedamount = 0.0;
         }
         return [
             'newHp' => max(0, $hp - $appliedamount),
@@ -181,12 +184,13 @@ class combat_engine {
     /**
      * Resolves one side's poison tick at the start of their turn.
      *
-     * @param int $hp Current HP of the side ticking.
+     * @param float $hp Current HP of the side ticking — see resolve_damage() for why this is
+     *  a float, not an int.
      * @param int $poisonrounds Rounds of poison damage still armed for that side.
      * @param float $tickdamage Damage dealt by a single tick (baseDamage).
      * @return array ['newHp' => float, 'newPoisonRounds' => int, 'ticked' => bool]
      */
-    public static function resolve_poison_tick(int $hp, int $poisonrounds, float $tickdamage): array {
+    public static function resolve_poison_tick(float $hp, int $poisonrounds, float $tickdamage): array {
         if ($poisonrounds <= 0) {
             return ['newHp' => $hp, 'newPoisonRounds' => $poisonrounds, 'ticked' => false];
         }
